# doc/openapi/swagger.yaml
basePath: /
definitions:
...
  model.CreateTeamRequest:
    properties:
      ID:
        type: string
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.TeamContent'
      metadata:
        $ref: '#/definitions/model.TeamMetadata'
    type: object
...
  model.ListTeamIdUuid:
    properties:
      teamIdUuids:
        items:
          $ref: '#/definitions/model.IdUuid'
        type: array
    type: object
  model.ListTeamsContentResponse:
    properties:
      teamContentResponses:
        items:
          $ref: '#/definitions/model.TeamContentResponse'
        type: array
    type: object
...
  model.Team:
    properties:
      ID:
        type: string
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.TeamContent'
      metadata:
        $ref: '#/definitions/model.TeamMetadata'
    type: object
...
  model.TeamMetadata:
    properties:
      dates:
        $ref: '#/definitions/model.CommonDate'
      name:
        type: string
    type: object
...
  model.UpdateTeamContentRequest:
    properties:
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.TeamContent'
    type: object
  model.UpdateTeamMetadataRequest:
    properties:
      UUID:
        type: string
      metadata:
        $ref: '#/definitions/model.TeamMetadata'
    type: object
  model.UpdateTeamRequest:
    properties:
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.TeamContent'
      metadata:
        $ref: '#/definitions/model.TeamMetadata'
    type: object
...
info:
  contact: {}
  description: This is a sample API server using Gin and Swagger.
  title: My API
  version: "1.0"
paths:
...
  /teams:
    get:
      description: Get Teams
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Team'
            type: array
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Teams
      tags:
      - Team
    post:
      consumes:
      - application/json
      description: Create Team
      parameters:
      - description: New team details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateTeamRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team created
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create Team
      tags:
      - Team
  /teams/{UUID}:
    delete:
      description: Delete Team by UUID
      parameters:
      - description: Team UUID
        in: path
        name: UUID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team deleted
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete Team by UUID
      tags:
      - Team
    get:
      description: Get Team by UUID
      parameters:
      - description: Team UUID
        in: path
        name: UUID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Team'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Team by UUID
      tags:
      - Team
    put:
      description: Update Team by UUID
      parameters:
      - description: Team UUID
        in: path
        name: UUID
        required: true
        type: string
      - description: Updated team details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTeamRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team updated
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update Team by UUID
      tags:
      - Team
  /teams/{UUID}/content:
    get:
      description: Get Team Content by UUID
      parameters:
      - description: Team UUID
        in: path
        name: UUID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team content found
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Team Content by UUID
      tags:
      - Team
    put:
      description: Update Team Content by UUID
      parameters:
      - description: Team UUID
        in: path
        name: UUID
        required: true
        type: string
      - description: Updated team content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTeamContentRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team content updated
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update Team Content by UUID
      tags:
      - Team
  /teams/{UUID}/metadata:
    get:
      description: Get Team Metadata by UUID
      parameters:
      - description: Team UUID
        in: path
        name: UUID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team metadata found
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Team Metadata by UUID
      tags:
      - Team
    put:
      description: Update Team Metadata by UUID
      parameters:
      - description: Team UUID
        in: path
        name: UUID
        required: true
        type: string
      - description: Updated team metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTeamMetadataRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team metadata updated
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update Team Metadata by UUID
      tags:
      - Team
  /teams/content:
    get:
      description: Get Teams Content
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListTeamsContentResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Teams Content
      tags:
      - Team
  /teams/email/{email}:
    get:
      description: Get Team by Email
      parameters:
      - description: Team email
        in: path
        name: email
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team found
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Team by Email
      tags:
      - Team
  /teams/id/{ID}:
    get:
      description: Get Team by ID
      parameters:
      - description: Team ID
        in: path
        name: ID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team found
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Team by ID
      tags:
      - Team
  /teams/ids-and-uuids:
    get:
      description: Get Team IDs and UUIDs
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListTeamIdUuid'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Team IDs and UUIDs
      tags:
      - Team
...
schemes:
- http
- https
swagger: "2.0"